embedding: centelize the build functions
